requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.2
Flask-Admin==1.6.0
greenlet==1.1.3.post0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
psycopg2-binary==2.9.4
python-dotenv==0.21.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.9.0
face-recognition==1.3.0
python-telegram-bot==13.14